// frontend/src/app/components/types.ts
export interface IUser {
    id: number;
    nome: string;
    cpf: string;
    dataNascimento: string;
    telefone: string;
    email: string;
    cep: string;
    endereco: string;
  }
  
  export interface IFormData {
    nome: string;
    cpf: string;
    dataNascimento: string;
    telefone: string;
    email: string;
    cep: string;
    endereco: string;
  }
  